% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsgn.R
\name{as.dsgn}
\alias{as.dsgn}
\title{Create and check for \code{dsgn} objects}
\usage{
as.dsgn(x)
}
\arguments{
\item{x}{A data.frame or coercible object with shifted time series columns labeled \code{VAR_tminus#}}
}
\value{
\itemize{
\item \code{as.dsgn()} returns a \code{dsgn} object
\item \code{is.dsgn()} returns TRUE if its argument is a design matrix object (that is, has \code{"dsgn"} amongst its classes) and FALSE otherwise
}
}
\description{
\code{dsgn} objects are recognized as design matrices with shifted, spatially distributed time series by other functions in the \code{stcd} package. Create \code{dsgn} objects using \code{as.dsgn()} and check if something is a \code{dsgn} object with \code{is.dsgn()}.
}
\examples{
design <- matrix(nrow = 2, ncol = 2)
is.dsgn(design)

# as.dsgn will throw an error if columns do not represent time series (i.e. if no column name contains `tminus`)
try(design <- as.dsgn(design))

colnames(design) <- c("var", "var_tminus1")
design <- as.dsgn(design)
is.dsgn(design)
}
