% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_neighborhood.R
\name{build_neighborhood}
\alias{build_neighborhood}
\title{Build neighborhood object from a dataframe of contiguity information}
\usage{
build_neighborhood(
  site_names,
  contiguity_df,
  site1 = "state1ab",
  site2 = "state2ab",
  contiguity_col = "conttype"
)
}
\arguments{
\item{site_names}{Vector of site names}

\item{contiguity_df}{data.frame with two columns for all site combinations
and a column indicating binary contiguity with 0/1}

\item{site1}{String indicating name of the column with site 1}

\item{site2}{String indicating name of the column with site 2}

\item{contiguity_col}{String indicating the name of the column with
contiguity information}
}
\value{
A named list of vectors, one for each site, which lists all its
neighboring sites

NOTE: For use in \link{fit_custom_gam}, all sites must have at least one neighbor
}
\description{
Build neighborhood object from a dataframe of contiguity information
}
\examples{

#Set up contiguity df for the boroughs of NYC
boroughs <- c("Manhattan", "The Bronx", "Queens", "Brooklyn", "Staten Island")
nyc_cont <- as.data.frame(t(combn(boroughs, 2, simplify = TRUE)))
nyc_cont$neighbor <- c(rep(0, 7), 1, rep(0, 2))

head(nyc_cont)

#Build neighborhood
nyc_nb <- build_neighborhood(site_names = boroughs,
              contiguity_df = nyc_cont,
              site1 = "V1",
              site2 = "V2",
              contiguity_col = "neighbor")

nyc_nb
}
