% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_independencies.R
\name{conditional_independencies}
\alias{conditional_independencies}
\title{Find conditional independencies of spatially-structured time series}
\usage{
conditional_independencies(
  data,
  max_lag = 2,
  sites = NULL,
  coords = NULL,
  time = "Year",
  family = "gaussian",
  modeltype = "separated",
  neighborhood = NULL,
  alpha = 0.05,
  collider_rule = "conservative",
  independence_test = "GCM",
  k0 = 7,
  nsim = 499,
  sufficiency = FALSE,
  silent = TRUE,
  skeleton_names = FALSE,
  skeleton_only = FALSE
)
}
\arguments{
\item{data}{Either a data.frame of time series and spatial information, or a
\code{dsgn} object as created by \link{shift_to_design}}

\item{max_lag}{Maximum time lag at which to test relationships}

\item{sites}{String naming column of site names, if in discrete space.
Otherwise, \code{NULL}.}

\item{coords}{Vector of strings naming columns with coordinate information,
if in coordinate space. Otherwise, \code{NULL}.}

\item{time}{String naming column with time information, if data.frame provided}

\item{family}{Ordered vector of strings indicating the distributions/family
of each time series. If one string is provided, it is assumed that is the
family of all variables.}

\item{modeltype}{One of \code{"separated"}, \code{"tensor"}, or \code{"both"}. See
\link{fit_custom_gam}.}

\item{neighborhood}{List defining neighborhood structure (only in discrete
space). See \link{build_neighborhood}.}

\item{alpha}{Significance level}

\item{collider_rule}{One of \code{"conservative"}, \code{"majority"}, or \code{"none"}. See
details.}

\item{independence_test}{Either \code{"GCM"} or \code{"WGCM"}. See \link{WGCM_fix}.}

\item{k0}{Number of quantiles for \link{WGCM_fix}.}

\item{nsim}{Number of simulations from which to create null distribution for
conditional independence tests.}

\item{sufficiency}{\code{TRUE} if all relevant variables observed. \code{FALSE} by
default.}

\item{silent}{If \code{FALSE}, reports more verbosely on function status. \code{TRUE} by
default.}

\item{skeleton_names}{If \code{TRUE}, returns a skeleton with nodes labeled by
variable name rather than index. \code{FALSE} by default. Do not set this to \code{TRUE}
if you plan to run the output through \link{orient_edges}.}

\item{skeleton_only}{If \code{TRUE}, interrupts the function before any collider
detection and returns the skeleton. \code{FALSE} by default.}
}
\value{
A list:
\describe{
\item{\code{skeleton}}{data.frame describing graph structure. \code{Xj} and \code{Xi}
indicate all pairs of nodes. \code{Xj_mark} and \code{Xi_mark} indicate the edge marks
at each end of the edge connecting the pair, or are \code{NA} if no edge exists.}
\item{\code{independencies}}{data.frame listing the independence tests run and
their results. Test statistics associated with tests of \code{Xi} is independent
of \code{Xj} given \code{S} are stored in the \code{Imin} column.}
\item{\code{triples}}{data.frame with triples \code{Xi_tau}--\code{Xk_t}--\code{Xj_t}. If
direction is yet to be resolved, \code{ambiguous} = 1. The \code{flag} column is
used internally to handle potential conflicts.}
\item{\code{names}}{vector of column names, if \code{skeleton_names = FALSE}}
\item{\code{sep_sets}}{data.frame listing, for all pairs of nodes \code{Xj}, \code{Xi},
the set conditional on which they are independent. The \code{S_sep} column
contains either the vector of indices comprising the separating set, \code{"null"}
if unconditionally independent, or \code{NA} if no such set was found.}
}
}
\description{
Function to find the conditional independencies and identify
colliders in a graph of time series with spatial structure. Heavily inspired
by the logic of PCMCI+ (Runge, 2020).
}
\section{Collider rules}{

To improve computational and statistical efficiency, PC-based algorithms
search for the minimum separating set for each pair of variables, rather than
for all separating sets. That is, the exclusion of a variable Xk from a
separating set Aij found in a PC-based algorithm does not in all cases
indicate that Xi ⊥ Xj | Xk, but just that a smaller set was found first.

Users can indicate how strictly they want to determine if Xk is a collider in
the structure Xi--Xk--Xj:
\itemize{
\item \code{collider_rule = "conservative"}: Xk is not in any separating sets. Default
for \code{conditional_independencies}.
\item \code{collider_rule = "majority"}: Xk is not in the majority of separating sets
\item \code{collider_rule = "none"}: Xk is not in the smallest/first found separating
set. Original rule and standard for PC-based algorithms.
}
}

\examples{

head(example_coord_data)
#reducing dimensions for example speed
smaller_data <- example_coord_data[,-c(1:5)]
head(smaller_data)


\dontrun{
example_inds <- conditional_independencies(data = smaller_data,
                  max_lag = 2,
                  coords = c("x", "y"),
                  family = c("gaussian", "gaussian", "betar"),
                  modeltype = "separated",
                  alpha = 0.025,
                  collider_rule = "none", #for example speed, not recommended
                  independence_test = "GCM",
                  sufficiency = FALSE #since we removed the Agr time series
                  )

str(example_inds)
}

}
