% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord_dsgn.R
\name{coord_dsgn}
\alias{coord_dsgn}
\alias{as.coord_dsgn}
\alias{is.coord_dsgn}
\title{Coerce to and check for \code{coord_dsgn} objects}
\usage{
as.coord_dsgn(x)

is.coord_dsgn(x)
}
\arguments{
\item{x}{A data.frame or coercible object with shifted time series columns labeled \code{VAR_tminus#}}
}
\value{
\itemize{
\item \code{as.coord_dsgn()} returns a \code{coord_dsgn} object
\item \code{is.coord_dsgn()} returns TRUE if its argument is a design matrix object (that is, has \code{"coord_dsgn"} amongst its classes) and FALSE otherwise
}
}
\description{
\code{coord_dsgn} objects are recognized as design matrices with shifted time series in coordinate/continuous space by other functions in the \code{stcd} package. Coerce objects to the \code{coord_dsgn} class using \code{as.dsgn()} and check if something is a \code{dsgn} object with \code{is.dsgn()}.

NOTE: To create a new \code{coord_dsgn} object from data with coordinates defined by two columns, use \code{\link[=shift_to_design]{shift_to_design()}} with a vector naming the two axes of coordinates for \code{coord}.
}
\examples{
design <- matrix(nrow = 2, ncol = 2)
is.coord_dsgn(design)

# as.coord_dsgn will throw an error if columns do not represent time series
# (i.e. if no column name contains `tminus`)
try(design <- as.coord_dsgn(design))

# it will not, however, throw any errors if you accidentally coerce something
# with a non-coordinate spatial structure.
# to create a new coord_dsgn object from data, see shift_to_design()
colnames(design) <- c("var", "var_tminus1")
design <- as.coord_dsgn(design)
is.coord_dsgn(design)

# coord_dsgn objects are also dsgn objects
is.dsgn(design)
}
\seealso{
\itemize{
\item \code{\link[=as.dsgn]{as.dsgn()}} for the umbrella class
\item \code{\link[=as.discrete_dsgn]{as.discrete_dsgn()}} for coercion to discrete-indexed spatial design matrices
\item \code{\link[=shift_to_design]{shift_to_design()}} for creation of new design matrices from data
}
}
