% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPDAG.R
\name{CPDAG}
\alias{CPDAG}
\alias{as.CPDAG}
\alias{is.CPDAG}
\title{CPDAG class -- WIP}
\usage{
as.CPDAG(x)

is.CPDAG(x)
}
\arguments{
\item{x}{A data frame representing nodes and edges in a time series complete
partially directed acyclic graph}
}
\value{
A \code{CPDAG} object
}
\description{
CPDAG class -- WIP
}
