% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete_dsgn.R
\name{discrete_dsgn}
\alias{discrete_dsgn}
\alias{as.discrete_dsgn}
\alias{is.discrete_dsgn}
\title{Coerce to and check for \code{discrete_dsgn} objects}
\usage{
as.discrete_dsgn(x)

is.discrete_dsgn(x)
}
\arguments{
\item{x}{A data.frame or coercible object with shifted time series columns labeled \code{VAR_tminus#}}
}
\value{
\itemize{
\item \code{as.discrete_dsgn()} returns a \code{discrete_dsgn} object
\item \code{is.discrete_dsgn()} returns TRUE if its argument is a design matrix object
(that is, has \code{"discrete_dsgn"} amongst its classes) and FALSE otherwise
}
}
\description{
\code{discrete_dsgn} objects are recognized as design matrices with shifted time series in discrete space by other functions in the \code{stcd} package. Coerce objects to the \code{discrete_dsgn} class using \code{as.dsgn()} and check if something is a \code{dsgn} object with \code{is.dsgn()}.

NOTE: To create a new \code{discrete_dsgn} object from data with a site ID column, use \code{\link[=shift_to_design]{shift_to_design()}} with a non-NULL value for \code{site_col}.
}
\examples{
design <- matrix(nrow = 2, ncol = 2)
is.discrete_dsgn(design)

# as.discrete_dsgn will throw an error if columns do not represent time series
# (i.e. if no column name contains `tminus`)
try(design <- as.discrete_dsgn(design))

# it will not, however, throw any errors if you accidentally coerce something
# with a non-discrete spatial structure.
# to create a new discrete_dsgn object from data, see shift_to_design()
colnames(design) <- c("var", "var_tminus1")
design <- as.discrete_dsgn(design)
is.discrete_dsgn(design)

# discrete_dsgn objects are also dsgn objects
is.dsgn(design)
}
\seealso{
\itemize{
\item \code{\link[=as.dsgn]{as.dsgn()}} for the umbrella class
\item \code{\link[=as.coord_dsgn]{as.coord_dsgn()}} for coercion to coordinate-indexed spatial design matrices
\item \code{\link[=shift_to_design]{shift_to_design()}} for creation of new design matrices from data
}
}
