% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_coord_data.R
\docType{data}
\name{example_coord_data}
\alias{example_coord_data}
\title{Example coordinate space design matrix}
\format{
\subsection{\code{example_coord_data}}{

A data frame with 5,226 rows and 22 columns:
\describe{
\item{Agr, Agr_tminus1, ..., Agr_tminus4}{\% of land used for agriculture}
\item{Fence, Fence_tminus1, ..., Fence_tminus4}{standardized amount of fencing}
\item{Fert, Fert_tminus1, ..., Fert_tminus4}{standardized amount of fertilizer used}
\item{Occ, Occ_tminus1, ..., Occ_tminus4}{probability of occupancy}
\item{x, y}{site coordinates, in the British National Grid CRS}
...
}
}
}
\usage{
example_coord_data
}
\description{
A simulated \code{coord_dsgn} object with shifted time series of agricultural
activity and probability of a species' occupancy at randomly sampled sites
across Great Britain.
}
\keyword{datasets}
