% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_disc_path.R
\name{find_disc_path}
\alias{find_disc_path}
\title{Find discriminating paths as part of edge orientation}
\usage{
find_disc_path(graph, i, k, j)
}
\arguments{
\item{graph}{Graph object as produced within \link{orient_edges}}

\item{i}{Indices of nodes to check for discriminating paths between}

\item{k}{Indices of nodes to check for discriminating paths between}

\item{j}{Indices of nodes to check for discriminating paths between}
}
\value{
A vector of node indices forming the discriminating path, or \code{NA} if
no path is found
}
\description{
Find discriminating paths as part of edge orientation
}
