% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_or_fit.R
\name{find_or_fit}
\alias{find_or_fit}
\title{Find or fit new gam if residuals are not already stored in temporary
environment}
\usage{
find_or_fit(
  design,
  response = a,
  model_on = b,
  distribution = distributions[a],
  Env = res_store,
  modeltype = modeltype,
  spatial_str = space
)
}
\arguments{
\item{design}{Design matrix}

\item{response}{Column index of response variable}

\item{model_on}{Vector of column index/indices of predictor variables, or
\code{"null"} for a space-only model}

\item{distribution}{String describing distribution/family of response}

\item{Env}{temporary environment storing residuals while running
\link{conditional_independencies}}

\item{modeltype}{String describing model structure: one of "separated",
"tensor", or "both". Not relevant when \code{model_on = "null"}.}

\item{spatial_str}{named list with either coords, or sites & nb}
}
\value{
residuals of the designated model, either via accessing the temporary
environment or by calling \link{fit_custom_gam}
}
\description{
Find or fit new gam if residuals are not already stored in temporary
environment
}
\keyword{internal}
