% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_custom_gam.R
\name{fit_custom_gam}
\alias{fit_custom_gam}
\title{Fit spatial generalized additive models for conditional independence testing}
\usage{
fit_custom_gam(
  design,
  response = a,
  model_on = b,
  distribution = distribution,
  modeltype = modeltype,
  ...
)
}
\arguments{
\item{design}{Design matrix}

\item{response}{Column index of response variable}

\item{model_on}{Vector of column index/indices of predictor variables, or
\code{"null"} for a space-only model}

\item{distribution}{String describing distribution/family of response}

\item{modeltype}{String describing model structure: one of "separated",
"tensor", "both", or "3D" (only for \code{coord_dsgn} objects). Not relevant when
\code{model_on = "null"}.}

\item{...}{other arguments specific to spatial structure. See details.}
}
\value{
Vector of raw residuals
}
\description{
Fit spatial generalized additive models for conditional independence testing
}
\section{Fitting discrete space models}{

When a \code{discrete_dsgn} object is passed to \code{fit_custom_gam}, it fits a
\link[mgcv:smooth.construct.mrf.smooth.spec]{Gaussian Markov random field}
which is either:
\itemize{
\item added to thin plate splines of variables indicated by the \code{model_on}
argument (\code{modeltype = "separated"})
\item in a tensor product with variables indicated by the \code{model_on}
argument (\code{modeltype = "tensor"})
\item both of the above (\code{modeltype = "both"})
}

The following arguments must also be provided for discrete space models:
\itemize{
\item \code{sites} String naming the column of \code{design} containing site names
\item \code{nb} List defining neighborhood structure. See the \link[mgcv:smooth.construct.mrf.smooth.spec]{smooth
specifications} for more details. See \link{build_neighborhood} to construct.
}
}

\section{Fitting coordinate space models}{

When a \code{coord_dsgn} object is passed to \code{fit_custom_gam}, it fits a
2-dimensional \link[mgcv:smooth.terms]{thin plate spline} (\code{bs = "tp"}), either additively,
in a tensor product, or both (see description of model types in above section).

The following argument must also be provided for coordinate space models:
\itemize{
\item \code{coords} A vector of strings naming the two columns containing coordinate
values in \code{design}
}
}

\examples{

#Provide a coord_dsgn object
head(example_coord_data)

#Fit spatial model of Agr ~ Agr_tminus1 + Agr_tminus2 and return residuals
resid_543 <- fit_custom_gam(design = example_coord_data,
                response = 1,
                model_on = c(2, 3),
                modeltype = "separated",
                distribution = "betar",
                coords = c("x", "y"))

resid_543

}
