% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orient_edges.R
\name{orient_edges}
\alias{orient_edges}
\title{Orient edges of a skeleton graph}
\usage{
orient_edges(independencies, max_lag = 2, names = TRUE, sufficiency = FALSE)
}
\arguments{
\item{independencies}{Outputs from conditional independence algorithm \link{conditional_independencies}}

\item{max_lag}{Maximum considered time lag}

\item{names}{If = TRUE, replace indices with variable names in output}

\item{sufficiency}{FALSE by default}
}
\value{
A PAG object if sufficiency = FALSE, a CPDAG if TRUE
}
\description{
Function to orient edges/place edge marks given a skeleton and identified colliders from
\link{conditional_independencies}. User indicates if sufficiency-assuming rules should be used or not, leading
to output of either a CPDAG or PAG.
}
\examples{

#' head(example_coord_data)
#reducing dimensions for example speed
smaller_data <- example_coord_data[,-c(1:5)]
head(smaller_data)


\dontrun{
example_inds <- conditional_independencies(data = smaller_data,
                  max_lag = 2,
                  coords = c("x", "y"),
                  family = c("gaussian", "gaussian", "betar"),
                  modeltype = "separated",
                  alpha = 0.025,
                  collider_rule = "none", #for example speed, not recommended
                  independence_test = "GCM",
                  sufficiency = FALSE #since we removed the Agr time series
                  )

str(example_inds)

example_or <- orient_edges(independencies = example_inds,
                max_lag = 2,
                names = TRUE,
                sufficiency = FALSE
                )

str(example_or)

}
}
