% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAG.R
\name{PAG}
\alias{PAG}
\alias{as.PAG}
\alias{is.PAG}
\title{PAG class -- WIP}
\usage{
as.PAG(x)

is.PAG(x)
}
\arguments{
\item{x}{A data frame representing nodes and edges in a time series partial ancestral graph}
}
\value{
A \code{PAG} object
}
\description{
PAG class -- WIP
}
