% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift_to_design.R
\name{shift_to_design}
\alias{shift_to_design}
\title{Create shifted time series design matrices from data}
\usage{
shift_to_design(data, time_col = "Time", max_lag = 2, ...)
}
\arguments{
\item{data}{\code{data.frame} or \code{list} of \link[terra:SpatRaster-class]{SpatRaster} objects
\itemize{
\item \code{data.frame} objects must have columns for each time series variable in the system, (a) column(s) for spatial structure, and a column indexing the time.
\item A \code{list} of \code{SpatRaster} objects must be in time order, with identically named data layers for each \code{SpatRaster} object
}}

\item{time_col}{String naming the column indexing time}

\item{max_lag}{Integer value representing the maximum time lag at which you expect relationships to exist. In practice, the function will shift time series by 2 times the \code{max_lag} for lagged relationship discovery.}

\item{...}{other arguments to specify spatial structure, see instructions below.}
}
\value{
a \code{dsgn} object (either \code{discrete_dsgn} or \code{coord_dsgn} depending on spatial structure)
}
\description{
Create shifted time series design matrices from data
}
\section{Creating discrete space design matrices}{


Time series indexed by discrete space (sites, counties, countries, etc) should be shifted to
discrete design matrices. Discrete space design matrices (objects of class \code{discrete_dsgn})
can only be created from \code{data.frame} objects. This is done by setting the \code{site_col} argument to
a string naming the column containing site information.
}

\section{Creating coordinate/continuous space design matrices}{


Time series indexed by coordinate/continuous space in 2 dimensions (x-y, latitude-longitude,
easting-northing, etc) should be shifted to coordinate design matrices. Coordinate space design
matrices (objects of class \code{coord_dsgn}) can be created from either a \code{data.frame} object or a
\code{list} of \link[terra:SpatRaster-class]{SpatRaster} objects:
\itemize{
\item Create a \code{coord_dsgn} object from a \code{data.frame} by setting the \code{coords} argument to a vector
of strings indicating the two column names containing coordinate information
(e.g. \code{coords = c("x", "y")}).
\item Creation of a design matrix from a \code{list} of \code{SpatRaster} objects will only ever create a
\code{coord_dsgn} object. You will need to define the number of samples to take from the rasters with
the \code{n_sample} argument (by default, \code{n_samples = 100}). This is a random sample, but you could set
a seed to ensure reproducibility.
}
}

\examples{
# The Chloris_chloris_data set has a discrete spatial structure with a single
# column designating the site ID (`CountryGroup`)
head(Chloris_chloris_data)

# Make sure to set the `site_col` argument to create a `discrete_dsgn` object
CC_design <- shift_to_design(data = Chloris_chloris_data,
                time_col = "Year",
                max_lag = 2,
                site_col = "CountryGroup")
head(CC_design)
class(CC_design)
}
