% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_system_linear.R
\name{sim_system_linear}
\alias{sim_system_linear}
\title{Simulate spatially-confounded systems of time series from a linear structural
causal model}
\usage{
sim_system_linear(
  mask,
  area_ID,
  t_steps,
  autocorrelation = 0.1,
  vgm_data = "Gau",
  nugget = 0,
  sill = 1,
  range_percent = 50,
  spat_conf = 1,
  k = 4,
  n_sample = 2500,
  strength = c("none", "low", "med", "high"),
  noise = 0.1,
  return_rast = FALSE,
  discrete = TRUE,
  n_sites = NULL
)
}
\arguments{
\item{mask}{SpatVector of polygons of countries/boundaries}

\item{area_ID}{Name of object that stores the area IDs in the mask}

\item{t_steps}{Number of time steps}

\item{autocorrelation}{Autocorrelation between years, rho}

\item{vgm_data}{Variogram model type for data - "Exp", "Sph", or "Gau"}

\item{nugget}{Variance at 0 distance}

\item{sill}{Asymptotic max variance}

\item{range_percent}{Percent of mask radius at which 95\% of max variance
reached}

\item{spat_conf}{Coefficient for effect of spatial confounding}

\item{k}{Neighbors to consider when simulating}

\item{n_sample}{Number of points to sample for simulation}

\item{strength}{Effect size strength}

\item{noise}{Variance of non-spatial random Gaussian noise with mean 0}

\item{return_rast}{If TRUE, return raw raster data}

\item{discrete}{Set = TRUE for discrete space, FALSE for continuous}

\item{n_sites}{Number of sites to simulate. Set to NULL to use number of
sub-polygons in mask.}
}
\value{
Data frame of time series at locations, and rasters if return_rast = TRUE
}
\description{
Simulate spatially-confounded systems of time series from a linear structural
causal model
}
