% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stcd-package.R
\docType{package}
\name{stcd-package}
\alias{stcd}
\alias{stcd-package}
\title{stcd: Spatiotemporal Causal Discovery}
\description{
Causal discovery algorithm for spatially distributed time series and functions to simulate and evaluate output.
}
\author{
\strong{Maintainer}: Rebecca Supple \email{rbs9@st-andrews.ac.uk} (\href{https://orcid.org/0000-0002-9752-0656}{ORCID})

Authors:
\itemize{
  \item Rebecca Supple \email{rbs9@st-andrews.ac.uk} (\href{https://orcid.org/0000-0002-9752-0656}{ORCID})
}

}
\keyword{internal}
