% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_graph.R
\name{t_graph}
\alias{t_graph}
\alias{as.t_graph}
\alias{is.t_graph}
\title{t_graph class -- WIP}
\usage{
as.t_graph(x)

is.t_graph(x)
}
\arguments{
\item{x}{A data frame representing nodes and edges in a time series graph}
}
\value{
A \code{t_graph} object
}
\description{
t_graph class -- WIP
}
