% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unpack_files.R
\name{unpack_files}
\alias{unpack_files}
\title{Unpack simulations from bulk files into the environment}
\usage{
unpack_files(
  type = "all",
  path,
  n_sims,
  n_reps,
  discrete = FALSE,
  sufficiency = TRUE
)
}
\arguments{
\item{type}{A vector of characters indicating what to unpack - any
combination of c("seed", "design", "ind", "or") or "all"}

\item{path}{String indicating path to folder where bulk files are kept}

\item{n_sims}{Number of simulated scenarios}

\item{n_reps}{Number of repetitions per scenario}

\item{discrete}{TRUE for discrete space, FALSE for continuous}

\item{sufficiency}{TRUE for sufficient simulations, FALSE for nonsufficient}
}
\value{
Adds to the environment lists of all the indicated files of the
indicated types
}
\description{
Unpack simulations from bulk files into the environment
}
