% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGCM_fix.R
\name{WGCM_fix}
\alias{WGCM_fix}
\title{Fixed weighted generalized covariance measure test}
\usage{
WGCM_fix(resid.XonZ, resid.YonZ, Z, k0, nsim = 499)
}
\arguments{
\item{resid.XonZ}{vector of raw residuals from a regression of X ~ Z}

\item{resid.YonZ}{vector of raw residuals from a regression of Y ~ Z}

\item{Z}{vector or matrix of condition variables}

\item{k0}{number of quantiles of Z over which to calculate test statistics}

\item{nsim}{number of simulations used to construct the null distribution}
}
\value{
A list containing the test statistic (\code{test.statistic}) and p-value
(\code{p.value}) of the wGCM test.
}
\description{
Tests against the null hypothesis that X is independent of Y given Z. Allows
for some nonlinearity in the dependence relationship, which is accommodated
by taking a weighted average of the test statistic across the distribution of
Z. This version allows for users to fit their own approximating regressions
outside of the wGCM test function. See Scheidegger, Hoerrmann and Buehlmann
(2022) "The Weighted Generalised Covariance Measure" for more details, and
the weightedGCM package for a fuller implementation that includes model
fitting.
}
\examples{
#Simulate X --> Z --> Y
X <- runif(100, -1, 1)
Z <- X^2 + rnorm(100, 0, 0.1)
Y <- exp(Z) + rnorm(100, 0, 0.1)

#X and Y should be independent conditional on Z
residXZ <- residuals(mgcv::gam(X ~ Z))
residYZ <- residuals(mgcv::gam(Y ~ Z))

test_result <- WGCM_fix(resid.XonZ = residXZ,
  resid.YonZ = residYZ,
  Z = Z,
  k0 = 10,
  nsim = 499)

test_result

}
