% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsgn.R
\name{dsgn}
\alias{dsgn}
\alias{as.dsgn}
\alias{is.dsgn}
\title{Coerce to and check for \code{dsgn} objects}
\usage{
as.dsgn(x)

is.dsgn(x)
}
\arguments{
\item{x}{A data.frame or coercible object with shifted time series columns labeled \code{VAR_tminus#}}
}
\value{
\itemize{
\item \code{as.dsgn()} returns a \code{dsgn} object
\item \code{is.dsgn()} returns TRUE if its argument is a design matrix object (that is,
has \code{"dsgn"} amongst its classes) and FALSE otherwise
}
}
\description{
\code{dsgn} objects are recognized as design matrices with shifted, spatially
distributed time series by other functions in the \code{stcd} package. Coerce objects
to the \code{dsgn} class using \code{as.dsgn()} and check if something is a \code{dsgn} object
with \code{is.dsgn()}.

NOTE: To create new \code{dsgn} objects from data, see \code{\link[=shift_to_design]{shift_to_design()}}
}
\examples{
design <- matrix(nrow = 2, ncol = 2)
is.dsgn(design)

# as.dsgn will throw an error if columns do not represent time series
# (i.e. if no column name contains `tminus`)
try(design <- as.dsgn(design))

colnames(design) <- c("var", "var_tminus1")
design <- as.dsgn(design)
is.dsgn(design)
}
\seealso{
\itemize{
\item \code{\link[=as.discrete_dsgn]{as.discrete_dsgn()}} and \code{\link[=as.coord_dsgn]{as.coord_dsgn()}} for the coercion to the subclasses of
discrete-indexed spatial design matrices and coordinate-indexed spatial design matrices,
respectively.
\item \code{\link[=shift_to_design]{shift_to_design()}} for creation of new design matrices from data
}
}
