% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsgn.R
\name{dsgn}
\alias{dsgn}
\title{Functions associated with 'dsgn' objects}
\usage{
dsgn(x)
}
\arguments{
\item{x}{A data.frame or coercible object}
}
\value{
A 'dsgn' object or TRUE/FALSE
}
\description{
Functions associated with 'dsgn' objects
}
\examples{
design <- matrix(nrow = 2, ncol = 2)
is.dsgn(design)

design <- as.dsgn(design)
is.dsgn(design)
}
