% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_field.R
\name{sim_field}
\alias{sim_field}
\title{Wrapper for gstat to simulate from a variogram model}
\usage{
sim_field(xy, vgm_model = "Gau", nugget, sill, range, k = 4, n_sample = 1000)
}
\arguments{
\item{xy}{Data frame of xy coordinates of point samples}

\item{vgm_model}{Variogram model type - "Exp", "Sph", or "Gau"}

\item{nugget}{Variance at 0 distance}

\item{sill}{Asymptotic max variance}

\item{range}{Distance at which 95\% of max variance reached (m)}

\item{k}{Number of nearest neighbors to consider when simulating}

\item{n_sample}{Number of points to sample for simulation}
}
\value{
A data frame of samples from the spatial model with values in a
column named "t1"
}
\description{
Wrapper for gstat to simulate from a variogram model
}
\examples{

points <- as.data.frame(unique(example_coord_data[,c("x", "y")]))
ex_field <- sim_field(xy = points,
             vgm_model = "Gau",
             nugget = 0,
             sill = 1,
             range = 10000,
             k = 4,
             n_sample = 1000)
head(ex_field)

}
